Widen CALL/CALL_NATIVE func_idx to 32-bit (two-word encoding)
The previous single-word encoding packed func_idx into operand b (8-bit),
capping each module at 256 functions and silently truncating in the
lowering pass via static_cast<u8>(it->second). Any module with more
than 256 functions would mis-call.

New encoding:
  Word 1: [op:8][dst:8][_:8][arg_count:8]
  Word 2: [func_idx:32]

Full 32-bit func_idx removes the ceiling outright; upper bits of word 2
are reserved for future inline-cache slots or tail-call flags. Also
widens CallData::native_index and IRBuilder::emit_call_native from u8
to u32 to fix the same silent truncation upstream of bytecode emission.

Wall-time impact on benchmarks: within noise. Quicksort (1.25M CALLs,
the only call-heavy workload of the three) pays one extra *pc++ per
call, measured at ~+1.6% — acceptable for removing a hard correctness
limit.

+++ include/roxy/compiler/ir_builder.hpp
@@ -82,7 +82,7 @@ private:
     ValueId try_simplify_unary(IROp op, ValueId operand, Type* result_type);
 
     ValueId emit_call(StringView func_name, Span<ValueId> args, Type* result_type);
-    ValueId emit_call_native(StringView func_name, Span<ValueId> args, Type* result_type, u8 native_index);
+    ValueId emit_call_native(StringView func_name, Span<ValueId> args, Type* result_type, u32 native_index);
     ValueId emit_call_external(StringView module_name, StringView func_name, Span<ValueId> args, Type* result_type);
     ValueId emit_index_get(ValueId container, ValueId index, ContainerKind kind, Type* result_type);
     void emit_index_set(ValueId container, ValueId index, ValueId value, ContainerKind kind);
